+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - Removed `es.regexp.flags` dependency from `es.regexp.to-string`, [#536](https://github.com/zloirock/core-js/issues/536), [#537](https://github.com/zloirock/core-js/issues/537)
 - Fixed IE8- non-enumerable properties support in `Object.{ assign, entries, values }`, [#541](https://github.com/zloirock/core-js/issues/541)
 - Fixed support of primitives in `Object.getOwnPropertySymbols` in Chrome 38 / 39, [#539](https://github.com/zloirock/core-js/issues/539)
-- `window.postMessage`-based task implementation uses `window.location.host` over `'*'`, [#542](https://github.com/zloirock/core-js/issues/542)
+- `window.postMessage`-based task implementation uses location origin over `'*'`, [#542](https://github.com/zloirock/core-js/issues/542)
 - Lookup `PromiseConstructor.resolve` only once in `Promise` combinators, [tc39/ecma262#1506](https://github.com/tc39/ecma262/pull/1506)
 - Temporarily removed `core-js` dependency from `core-js-compat` since it's required for missed at this moment feature
 - Added compat data for Chrome 76, FF 67, Node 12